blockcache: Handle a block being dereferenced that's already in the LRU
If we got two requests for the same block, they'll both decompress
without holding any lock. Later, when the second one gets dereferenced,
it'll try to add it to the LRU, but that doesn't trigger an eviction of
the one already in the cache. If using the s2malloc allocator, that'll
cause us to leak memory.